+++ cvat-ui/src/components/create-task-page/create-task-page.tsx
@@ -4,6 +4,7 @@
 
 import './styles.scss';
 import React, { useEffect } from 'react';
+import { useLocation } from 'react-router';
 import { Row, Col } from 'antd/lib/grid';
 import Modal from 'antd/lib/modal';
 import Text from 'antd/lib/typography/Text';
@@ -21,7 +22,17 @@ interface Props {
 }
 
 export default function CreateTaskPage(props: Props): JSX.Element {
-    const { error, status, taskId, onCreate, installedGit } = props;
+    const {
+        error, status, taskId, onCreate, installedGit,
+    } = props;
+
+    const location = useLocation();
+
+    let projectId = null;
+    const params = new URLSearchParams(location.search);
+    if (params.get('projectId')?.match(/^[1-9]+[0-9]*$/)) {
+        projectId = +(params.get('projectId') as string);
+    }
 
     useEffect(() => {
         if (error) {
@@ -61,7 +72,13 @@ export default function CreateTaskPage(props: Props): JSX.Element {
         <Row type='flex' justify='center' align='top' className='cvat-create-task-form-wrapper'>
             <Col md={20} lg={16} xl={14} xxl={9}>
                 <Text className='cvat-title'>Create a new task</Text>
-                <CreateTaskContent taskId={taskId} status={status} onCreate={onCreate} installedGit={installedGit} />
+                <CreateTaskContent
+                    taskId={taskId}
+                    projectId={projectId}
+                    status={status}
+                    onCreate={onCreate}
+                    installedGit={installedGit}
+                />
             </Col>
         </Row>
     );