+++ cvat-ui/src/components/task-page/details.tsx
@@ -7,18 +7,17 @@ import { Row, Col } from 'antd/lib/grid';
 import Tag from 'antd/lib/tag';
 import Icon from 'antd/lib/icon';
 import Modal from 'antd/lib/modal';
-import Button from 'antd/lib/button';
 import notification from 'antd/lib/notification';
 import Text from 'antd/lib/typography/Text';
 import Title from 'antd/lib/typography/Title';
 import moment from 'moment';
 
 import getCore from 'cvat-core-wrapper';
-import patterns from 'utils/validation-patterns';
 import { getReposData, syncRepos } from 'utils/git-utils';
 import { ActiveInference } from 'reducers/interfaces';
 import AutomaticAnnotationProgress from 'components/tasks-page/automatic-annotation-progress';
 import UserSelector, { User } from './user-selector';
+import BugTrackerEditor from './bug-tracker-editor';
 import LabelsEditorComponent from '../labels-editor/labels-editor';
 
 const core = getCore();
@@ -34,8 +33,6 @@ interface Props {
 
 interface State {
     name: string;
-    bugTracker: string;
-    bugTrackerEditing: boolean;
     repository: string;
     repositoryStatus: string;
 }
@@ -57,8 +54,6 @@ export default class DetailsComponent extends React.PureComponent<Props, State> 
         this.previewWrapperRef = React.createRef<HTMLDivElement>();
         this.state = {
             name: taskInstance.name,
-            bugTracker: taskInstance.bugTracker,
-            bugTrackerEditing: false,
             repository: '',
             repositoryStatus: '',
         };
@@ -119,7 +114,6 @@ export default class DetailsComponent extends React.PureComponent<Props, State> 
         if (prevProps !== this.props) {
             this.setState({
                 name: taskInstance.name,
-                bugTracker: taskInstance.bugTracker,
             });
         }
     }
@@ -194,7 +188,7 @@ export default class DetailsComponent extends React.PureComponent<Props, State> 
         );
     }
 
-    private renderUsers(): JSX.Element {
+    private renderDescription(): JSX.Element {
         const { taskInstance, onTaskUpdate } = this.props;
         const owner = taskInstance.owner ? taskInstance.owner.username : null;
         const assignee = taskInstance.assignee ? taskInstance.assignee : null;
@@ -211,7 +205,11 @@ export default class DetailsComponent extends React.PureComponent<Props, State> 
 
         return (
             <Row className='cvat-task-details-user-block' type='flex' justify='space-between' align='middle'>
-                <Col span={12}>{owner && <Text type='secondary'>{`Created by ${owner} on ${created}`}</Text>}</Col>
+                <Col span={12}>
+                    {owner && (
+                        <Text type='secondary'>{`Task #${taskInstance.id} Created by ${owner} on ${created}`}</Text>
+                    )}
+                </Col>
                 <Col span={10}>
                     <Text type='secondary'>Assigned to</Text>
                     {assigneeSelect}
@@ -294,86 +292,6 @@ export default class DetailsComponent extends React.PureComponent<Props, State> 
         );
     }
 
-    private renderBugTracker(): JSX.Element {
-        const { taskInstance, onTaskUpdate } = this.props;
-        const { bugTracker, bugTrackerEditing } = this.state;
-
-        let shown = false;
-        const onStart = (): void => {
-            this.setState({
-                bugTrackerEditing: true,
-            });
-        };
-        const onChangeValue = (value: string): void => {
-            if (value && !patterns.validateURL.pattern.test(value)) {
-                if (!shown) {
-                    Modal.error({
-                        title: `Could not update the task ${taskInstance.id}`,
-                        content: 'Issue tracker is expected to be URL',
-                        onOk: () => {
-                            shown = false;
-                        },
-                    });
-                    shown = true;
-                }
-            } else {
-                this.setState({
-                    bugTracker: value,
-                    bugTrackerEditing: false,
-                });
-
-                taskInstance.bugTracker = value;
-                onTaskUpdate(taskInstance);
-            }
-        };
-
-        if (bugTracker) {
-            return (
-                <Row>
-                    <Col>
-                        <Text strong className='cvat-text-color'>
-                            Issue Tracker
-                        </Text>
-                        <br />
-                        <Text editable={{ onChange: onChangeValue }}>{bugTracker}</Text>
-                        <Button
-                            type='ghost'
-                            size='small'
-                            onClick={(): void => {
-                                // false positive
-                                // eslint-disable-next-line
-                                window.open(bugTracker, '_blank');
-                            }}
-                            className='cvat-open-bug-tracker-button'
-                        >
-                            Open the issue
-                        </Button>
-                    </Col>
-                </Row>
-            );
-        }
-
-        return (
-            <Row>
-                <Col>
-                    <Text strong className='cvat-text-color'>
-                        Issue Tracker
-                    </Text>
-                    <br />
-                    <Text
-                        editable={{
-                            editing: bugTrackerEditing,
-                            onStart,
-                            onChange: onChangeValue,
-                        }}
-                    >
-                        {bugTrackerEditing ? '' : 'Not specified'}
-                    </Text>
-                </Col>
-            </Row>
-        );
-    }
-
     private renderLabelsEditor(): JSX.Element {
         const { taskInstance, onTaskUpdate } = this.props;
 
@@ -393,7 +311,10 @@ export default class DetailsComponent extends React.PureComponent<Props, State> 
     }
 
     public render(): JSX.Element {
-        const { activeInference, cancelAutoAnnotation } = this.props;
+        const {
+            activeInference, cancelAutoAnnotation, taskInstance, onTaskUpdate,
+        } = this.props;
+
         return (
             <div className='cvat-task-details'>
                 <Row type='flex' justify='start' align='middle'>
@@ -409,9 +330,17 @@ export default class DetailsComponent extends React.PureComponent<Props, State> 
                         </Row>
                     </Col>
                     <Col md={16} lg={17} xl={17} xxl={18}>
-                        {this.renderUsers()}
+                        {this.renderDescription()}
                         <Row type='flex' justify='space-between' align='middle'>
-                            <Col span={12}>{this.renderBugTracker()}</Col>
+                            <Col span={12}>
+                                <BugTrackerEditor
+                                    instance={taskInstance}
+                                    onChange={(bugTracker) => {
+                                        taskInstance.bugTracker = bugTracker;
+                                        onTaskUpdate(taskInstance);
+                                    }}
+                                />
+                            </Col>
                             <Col span={10}>
                                 <AutomaticAnnotationProgress
                                     activeInference={activeInference}
@@ -420,7 +349,7 @@ export default class DetailsComponent extends React.PureComponent<Props, State> 
                             </Col>
                         </Row>
                         {this.renderDatasetRepository()}
-                        {this.renderLabelsEditor()}
+                        {!taskInstance.projectId && this.renderLabelsEditor()}
                     </Col>
                 </Row>
             </div>