+++ cvat-ui/src/components/task-page/top-bar.tsx
@@ -3,6 +3,7 @@
 // SPDX-License-Identifier: MIT
 
 import React from 'react';
+import { useHistory } from 'react-router';
 import { Row, Col } from 'antd/lib/grid';
 import Button from 'antd/lib/button';
 import Dropdown from 'antd/lib/dropdown';
@@ -18,12 +19,27 @@ interface DetailsComponentProps {
 
 export default function DetailsComponent(props: DetailsComponentProps): JSX.Element {
     const { taskInstance } = props;
-    const { id } = taskInstance;
+
+    const history = useHistory();
 
     return (
         <Row className='cvat-task-top-bar' type='flex' justify='space-between' align='middle'>
             <Col>
-                <Text className='cvat-title'>{`Task details #${id}`}</Text>
+                {taskInstance.projectId ? (
+                    <Button
+                        onClick={() => history.push(`/projects/${taskInstance.projectId}`)}
+                        type='link'
+                        size='large'
+                    >
+                        <Icon type='left' />
+                        Back to project
+                    </Button>
+                ) : (
+                    <Button onClick={() => history.push('/tasks')} type='link' size='large'>
+                        <Icon type='left' />
+                        Back to tasks
+                    </Button>
+                )}
             </Col>
             <Col>
                 <Dropdown overlay={<ActionsMenuContainer taskInstance={taskInstance} />}>