+++ cvat-ui/src/reducers/root-reducer.ts
@@ -4,6 +4,7 @@
 
 import { combineReducers, Reducer } from 'redux';
 import authReducer from './auth-reducer';
+import projectsReducer from './projects-reducer';
 import tasksReducer from './tasks-reducer';
 import aboutReducer from './about-reducer';
 import shareReducer from './share-reducer';
@@ -20,6 +21,7 @@ import reviewReducer from './review-reducer';
 export default function createRootReducer(): Reducer {
     return combineReducers({
         auth: authReducer,
+        projects: projectsReducer,
         tasks: tasksReducer,
         about: aboutReducer,
         share: shareReducer,